Move coordinates to units
