+++ src/providers/units.rs
@@ -1,5 +1,6 @@
+use rocket::serde::Serialize;
 use serde::Deserialize;
-use std::fmt::Debug;
+use std::fmt::{Debug, Display, Formatter};
 
 #[derive(Deserialize, Debug, Copy, Clone)]
 pub struct Kelvin(f32);
@@ -71,4 +72,28 @@ impl Ratio {
             Self::Percentage(v) => f64::from(*v) / 100.0,
         }
     }
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct Coordinate(f32);
+
+impl Display for Coordinate {
+    // Standardize 7 digits for coordinates and that should be plenty
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{:.7}", self.0)
+    }
+}
+
+impl From<f32> for Coordinate {
+    fn from(value: f32) -> Self {
+        Self(value)
+    }
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct Coordinates {
+    #[serde(alias = "lat")]
+    pub latitude: Coordinate,
+    #[serde(alias = "lon")]
+    pub longitude: Coordinate,
 }
